refactor(rescue-3c): lift DaemonLockGuard into julie-core leader_lock (T2)
Move DaemonLockGuard, LockAlreadyHeld, AcquireError, HELD_DAEMON_LOCKS,
and the private helpers (held_daemon_locks, normalize_lock_path,
is_lock_already_held) from src/daemon/discovery.rs into the new
crates/julie-core/src/workspace/leader_lock.rs module.

Re-export all three public types from the old path so every existing
caller (app.rs, lock_test.rs, app_test.rs) compiles unchanged.

fs2 = "0.4" was already present in julie-core/Cargo.toml; no dep change.

TDD: RED confirmed (unresolved imports), GREEN on both julie-core inline
test (test_leader_lock_in_process_dedup) and the existing julie-crate test
(test_daemon_lock_try_acquire_fails_when_held). cargo check -p julie clean.
