+++ .memories/2026-06-05/005239_93a0.md
@@ -1,0 +1,32 @@
+---
+id: checkpoint_93a0b4c2
+timestamp: 2026-06-05T00:52:39.129Z
+tags:
+  - julie-rescue
+  - phase3c
+  - pr-3c1
+  - subagent-driven
+git:
+  branch: julie-rescue-phase3c
+  commit: d33e2980
+  files:
+    - .codex/config.toml
+summary: Julie Rescue PR 3c.1 — execution started
+briefId: julie-rescue-de-slop-packaging-keep-the-moat
+type: checkpoint
+next: Await core-dev (T1) + lock-dev (T2) reports → inline review each →
+  dispatch T13, then T3, then T4.
+---
+
+## Julie Rescue PR 3c.1 — execution started
+
+Phase 3c plan **approved** (codex pre-merge review per sub-PR). Executing PR **3c.1** (leader-lock + hang-guard infra) via razorback subagent-driven-development + TeamCreate visible team `julie-3c1` (Opus lead + Sonnet workers).
+
+- **Branch:** `julie-rescue-phase3c` (off merged main). Plan committed `d33e2980` (`docs/plans/2026-06-05-julie-phase3c-inprocess-leader.md`).
+- **Tasks:** T1 (julie-core `workspace_leader_lock` path helper), T2 (lift `DaemonLockGuard`→julie-core `workspace/leader_lock.rs` + re-export), T3 (handler per-request `tokio::time::timeout`), T4 (`new_in_process` ctor + `LeadershipState`, blockedBy T2+T3), T13 (dashboard persist-vs-drop doc table).
+- **Dispatch shape:** `src/handler.rs` touched by T3+T4 → SERIALIZE T3→T4. T2 edits `src/lib.rs` → land green before handler tasks compile-test. Round 1 (running NOW, parallel/disjoint): T1=`core-dev`, T2=`lock-dev`. Round 1.5: T13. Round 2: T3. Round 3: T4.
+
+### Discipline (carry-over lessons)
+- FREEZE every worker after its work is accepted (3b churn lesson [[feedback_freeze_all_workers_after_acceptance]]); idle ping ≠ freeze ack — verify via mtime.
+- Workers run ONLY narrow named tests (`cargo nextest run -p <crate> <test>`, ≤2 runs); lead owns `cargo xtask test dev` branch-gate.
+- NEVER commit `.codex/config.toml` / `.miller/`; ALWAYS commit `.memories/`.
