+++ .memories/2026-06-05/010120_33a2.md
@@ -1,0 +1,69 @@
+---
+id: checkpoint_33a2e3dc
+timestamp: 2026-06-05T01:01:20.393Z
+tags:
+  - phase3c
+  - leader-lock
+  - refactor
+  - julie-core
+  - DaemonLockGuard
+git:
+  branch: julie-rescue-phase3c
+  commit: be2855a0
+  files:
+    - .codex/config.toml
+    - crates/julie-core/src/workspace/leader_lock.rs
+    - crates/julie-core/src/workspace/mod.rs
+    - src/daemon/discovery.rs
+summary: T2 — Lifted DaemonLockGuard into julie-core leader_lock module
+briefId: julie-rescue-de-slop-packaging-keep-the-moat
+type: checkpoint
+evidence:
+  - cargo nextest run -p julie-core test_leader_lock_in_process_dedup → PASS
+  - cargo nextest run -p julie test_daemon_lock_try_acquire_fails_when_held →
+    PASS
+  - cargo check -p julie → clean
+symbols:
+  - DaemonLockGuard
+  - AcquireError
+  - LockAlreadyHeld
+  - leader_lock
+  - HELD_DAEMON_LOCKS
+next: "T3: Handler-level per-request deadline (in-process hang guard)"
+---
+
+## T2 — Lifted DaemonLockGuard into julie-core leader_lock module
+
+### WHAT
+Relocated the OS-advisory-lock primitive from `src/daemon/discovery.rs` into a new
+`crates/julie-core/src/workspace/leader_lock.rs` module. Added re-exports at the old
+path so all existing callers compile unchanged.
+
+### Items moved
+- `DaemonLockGuard` struct + `try_acquire` + `path` + `Drop` impl
+- `LockAlreadyHeld` struct + `Display` + `Error` impls
+- `AcquireError` enum + `From<LockAlreadyHeld>` + `Display` + `Error::source`
+- `HELD_DAEMON_LOCKS` static, `held_daemon_locks()`, `normalize_lock_path()`, `is_lock_already_held()`
+
+### HOW
+- `crates/julie-core/src/workspace/leader_lock.rs` — new module with full implementation + inline `#[cfg(test)] mod tests`
+- `crates/julie-core/src/workspace/mod.rs` — added `pub mod leader_lock;`
+- `src/daemon/discovery.rs` — removed lock code, added `pub use julie_core::workspace::leader_lock::{AcquireError, DaemonLockGuard, LockAlreadyHeld};`
+- `fs2 = "0.4"` was **already present** in julie-core/Cargo.toml — no change needed
+
+### Callers verified (all resolve via re-export)
+- `src/daemon/app.rs:24` — `use super::discovery::{AcquireError, DaemonLockGuard};`
+- `src/tests/daemon/lock_test.rs:24` — `use crate::daemon::discovery::{AcquireError, DaemonLockGuard, LockAlreadyHeld};`
+- `src/tests/daemon/app_test.rs:13` — `use crate::daemon::discovery::{AcquireError, DaemonLockGuard, DiscoveryFile, DiscoveryState};`
+- `src/daemon/legacy_migration.rs` / `singleton.rs` — use `SingletonLock::try_acquire` (different type, not in scope)
+
+### IMPACT
+- `DaemonLockGuard` now usable from non-daemon code for per-workspace leader election (Phase 3c goal)
+- Zero behavior change — pure relocation + re-export
+- All existing callers compile unchanged through the re-export
+
+### Evidence
+- RED: `cargo nextest run -p julie-core test_leader_lock_in_process_dedup` → compile error (types not defined)
+- GREEN (julie-core): `cargo nextest run -p julie-core test_leader_lock_in_process_dedup` → PASS 0.011s
+- GREEN (julie): `cargo nextest run -p julie test_daemon_lock_try_acquire_fails_when_held` → PASS 0.016s
+- `cargo check -p julie` → Finished clean
