+++ crates/julie-core/src/workspace/mod.rs
@@ -2,6 +2,7 @@
 //!
 //! These are zero-dependency utilities shared across the workspace stack.
 
+pub mod leader_lock;
 pub mod mutation_gate;
 pub mod registry;
 pub mod root_safety;
